added types, renamed components
