+++ app/snippets/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import SnippetGallery from "../components/SnippetsGalleryComponent";
+import SnippetGallery from "../components/SnippetsGallery";
 
 const Snippets: React.FC = () => {
   return (
